04_Code/modern_classifier/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers
transformers>=4.30.0
timm>=0.9.0  # PyTorch Image Models (for ViT variants)

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking

# Evaluation
scikit-learn>=1.2.0
